move Location from rewriter.h to types.h
